requirements.txt
Flask
flask_login
flask_sqlalchemy
mongoengine
flask_inject
python_dotenv
werkzeug